requirements.txt
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==11.0.0
django-filter==24.3
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
mysqlclient==2.2.6
python-decouple==3.8
celery==5.4.0
redis==5.2.1